set up website filesystem
Added multiple directories to organize the website. /src holds the pages
folder, which will hold all webpages, the assets folder to be
used for various reusable images later, the components folder which will
be used to divide reusable website components into small & manageable
chunks, and other various frontend aspects like the theme.tsx file,
which defines the standard of layout and styling for the website.

The /public folder exists for now mostly to hold images. The /articles
folder will house further documentation when it is added. Various other
directories and files have been added to ensure the website can be
developed on properly. Documentation will be written relating to those
things in the near future.